Add files via upload
Here, we have stated the steps with R code chunks that are required to generate a simulated data set of an individual randomized controlled trial. In addition, we have mentioned the steps along with R codes that helped us to perform 20 simulations at varying proportions of compliance ranging from 2 to 99%. At the end, we have provided the results of 20 simulations for the parameters of interest including true causal effect, average causal effect or ITT, CACE, and IV estimand. There is no difference between CACE and IV estimand, but the purpose of providing IV estimand is to demonstrate that in practice, clinicians and researchers can easily compute CACE using IVreg package in R. IVreg package with its functions help to estimate CACE or IV.